Add --verbose flag to Dockerfile
